+++ services/school/common/teacher/teacher_service.go
@@ -44,9 +44,15 @@ const DEFAULT_ERROR_MESSAGE = "interact with teacher model"
 const uidAcceptedLetters = "T"
 
 func (service *Service) Create(
-	inputJwtToken *types.JwtToken,
+	ctxData *types.ContextData,
 	request *data.TeacherRequest,
 ) (result *model.Teacher, errCode int, err error) {
+	// Check school
+	newRequest := *request
+	if ctxData.Jwt.SchoolID > 0 {
+		newRequest.SchoolID = ctxData.Jwt.SchoolID
+	}
+
 	// Get role
 	userRole, errRole := service.RoleService.Repository.GetByName(config.Env.FixtureRoleTeacher)
 	if errRole != nil || userRole == nil || userRole.ID < 1 {
@@ -56,7 +62,7 @@ func (service *Service) Create(
 	}
 
 	// Get the school
-	foundSchool, errSchool := service.SchoolService.Repository.GetByID(request.SchoolID)
+	foundSchool, errSchool := service.SchoolService.Repository.GetByID(newRequest.SchoolID)
 	if errSchool != nil {
 		errCode = http.StatusInternalServerError
 		err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
@@ -69,45 +75,45 @@ func (service *Service) Create(
 	}
 
 	// Check request
-	if request == nil || request.Info == nil {
+	if request == nil || newRequest.Info == nil {
 		errCode = http.StatusBadRequest
 		err = constants.Http400BadRequestErrorMessage()
 		return
 	}
 
 	// Generate the email
-	newEmail := request.Email
-	if request.AutoGenerateEmail {
+	newEmail := newRequest.Email
+	if newRequest.AutoGenerateEmail {
 		if foundSchool.Config == nil {
 			errCode = http.StatusNotFound
 			err = constants.Http404ErrorMessage("school")
 			return
 		}
 		newEmail = helpers.GenerateEmailFromFullName(
-			request.Info.FirstName,
-			request.Info.LastName,
+			newRequest.Info.FirstName,
+			newRequest.Info.LastName,
 			foundSchool.Config.UserEmailDomainName,
 		)
 	}
 
 	// Format request
 	var item = &dataUser.UserRequest{
-		SchoolID:    request.SchoolID,
+		SchoolID:    newRequest.SchoolID,
 		RoleID:      userRole.ID,
 		Email:       newEmail,
-		PhoneNumber: request.PhoneNumber,
+		PhoneNumber: newRequest.PhoneNumber,
 		IsActivated: true,
-		Status:      request.Status,
+		Status:      newRequest.Status,
 		Info: &dataUser.UserInfoRequest{
-			Gender:        request.Info.Gender,
-			Username:      request.Info.Username,
-			FirstName:     request.Info.FirstName,
-			LastName:      request.Info.LastName,
-			Birthday:      request.Info.Birthday,
-			BirthLocation: request.Info.BirthLocation,
-			Address:       request.Info.Address,
-			Language:      request.Info.Language,
-			Image:         request.Info.Image,
+			Gender:        newRequest.Info.Gender,
+			Username:      newRequest.Info.Username,
+			FirstName:     newRequest.Info.FirstName,
+			LastName:      newRequest.Info.LastName,
+			Birthday:      newRequest.Info.Birthday,
+			BirthLocation: newRequest.Info.BirthLocation,
+			Address:       newRequest.Info.Address,
+			Language:      newRequest.Info.Language,
+			Image:         newRequest.Info.Image,
 		},
 	}
 
@@ -132,10 +138,10 @@ func (service *Service) Create(
 	}
 
 	// Generate the uid if it is empty
-	var newUID = request.UID
+	var newUID = newRequest.UID
 	if len(newUID) < 1 {
 		// Get all uids to exclude
-		uids, errUids := service.Repository.GetAll(nil, nil, &data.GetAllRequest{SchoolID: request.SchoolID})
+		uids, errUids := service.Repository.GetAll(nil, nil, &data.GetAllRequest{SchoolID: newRequest.SchoolID})
 		if errUids != nil {
 			errCode = http.StatusInternalServerError
 			err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
@@ -161,7 +167,7 @@ func (service *Service) Create(
 
 	// Insert
 	result, err = service.Repository.Create(&model.Teacher{
-		SchoolID: request.SchoolID,
+		SchoolID: newRequest.SchoolID,
 		UserID:   createdUser.ID,
 		UID:      newUID,
 	})
@@ -174,15 +180,16 @@ func (service *Service) Create(
 }
 
 func (service *Service) CreateTeacherClassSubjectUnit(
-	inputJwtToken *types.JwtToken,
+	ctxData *types.ContextData,
 	request *data.TeacherClassSubjectUnitRequest,
 ) (result *model.TeacherClassSubjectUnit, errCode int, err error) {
 	// Format request
+	newRequest := *request
 	item := &model.TeacherClassSubjectUnit{
-		TeacherID:      request.TeacherID,
-		YearID:         request.YearID,
-		ClassSubjectID: request.ClassSubjectID,
-		UnitID:         request.UnitID,
+		TeacherID:      newRequest.TeacherID,
+		YearID:         newRequest.YearID,
+		ClassSubjectID: newRequest.ClassSubjectID,
+		UnitID:         newRequest.UnitID,
 	}
 
 	// Check unique
@@ -209,12 +216,17 @@ func (service *Service) CreateTeacherClassSubjectUnit(
 }
 
 func (service *Service) Update(
-	inputJwtToken *types.JwtToken,
+	ctxData *types.ContextData,
 	id int64,
 	request *data.TeacherRequest,
 ) (result *model.Teacher, errCode int, err error) {
-	// Check if exists
-	foundItem, err := service.Repository.GetByID(id)
+	// Check if the item exists
+	var foundItem *model.Teacher
+	if ctxData.User.Feature != constants.FeatureAdmin {
+		foundItem, err = service.Repository.GetByIDSchoolID(id, ctxData.Jwt.SchoolID)
+	} else {
+		foundItem, err = service.Repository.GetByID(id)
+	}
 	if err != nil {
 		errCode = http.StatusInternalServerError
 		err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
@@ -226,11 +238,17 @@ func (service *Service) Update(
 		return
 	}
 
+	// Check school
+	newRequest := *request
+	if ctxData.Jwt.SchoolID > 0 {
+		newRequest.SchoolID = ctxData.Jwt.SchoolID
+	}
+
 	// Generate the uid if it is empty
-	var newUID = request.UID
+	var newUID = newRequest.UID
 	if len(newUID) < 1 {
 		// Get all uids to exclude
-		uids, errUids := service.Repository.GetAll(nil, nil, &data.GetAllRequest{SchoolID: request.SchoolID})
+		uids, errUids := service.Repository.GetAll(nil, nil, &data.GetAllRequest{SchoolID: newRequest.SchoolID})
 		if errUids != nil {
 			errCode = http.StatusInternalServerError
 			err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
@@ -289,22 +307,22 @@ func (service *Service) Update(
 
 	// Update user
 	userRequest := dataUser.UserRequest{
-		SchoolID:    request.SchoolID,
+		SchoolID:    newRequest.SchoolID,
 		RoleID:      teacherRole.ID,
-		Email:       request.Email,
-		PhoneNumber: request.PhoneNumber,
+		Email:       newRequest.Email,
+		PhoneNumber: newRequest.PhoneNumber,
 		IsActivated: true,
-		Status:      request.Status,
+		Status:      newRequest.Status,
 		Info: &dataUser.UserInfoRequest{
-			Gender:        request.Info.Gender,
-			Username:      request.Info.Username,
-			FirstName:     request.Info.FirstName,
-			LastName:      request.Info.LastName,
-			Birthday:      request.Info.Birthday,
-			BirthLocation: request.Info.BirthLocation,
-			Address:       request.Info.Address,
-			Language:      request.Info.Language,
-			Image:         request.Info.Image,
+			Gender:        newRequest.Info.Gender,
+			Username:      newRequest.Info.Username,
+			FirstName:     newRequest.Info.FirstName,
+			LastName:      newRequest.Info.LastName,
+			Birthday:      newRequest.Info.Birthday,
+			BirthLocation: newRequest.Info.BirthLocation,
+			Address:       newRequest.Info.Address,
+			Language:      newRequest.Info.Language,
+			Image:         newRequest.Info.Image,
 		},
 	}
 	_, errCodeUser, errUser := service.UserService.Update(nil, foundItem.UserID, &userRequest)
@@ -329,29 +347,40 @@ func (service *Service) Update(
 }
 
 func (service *Service) UpdateTeacherClassSubjectUnit(
-	inputJwtToken *types.JwtToken,
+	ctxData *types.ContextData,
 	id int64,
 	request *data.TeacherClassSubjectUnitRequest,
 ) (result *model.TeacherClassSubjectUnit, errCode int, err error) {
-	// Check if exists
-	foundItem, err := service.Repository.GetTeacherClassSubjectUnitByID(id)
+	// Check if the item exists
+	var foundItem *model.TeacherClassSubjectUnit
+	if ctxData.User.Feature != constants.FeatureAdmin {
+		foundItem, err = service.Repository.GetByIDSchoolID(id, ctxData.Jwt.SchoolID)
+	} else {
+		foundItem, err = service.Repository.GetByID(id)
+	}
 	if err != nil {
 		errCode = http.StatusInternalServerError
 		err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
 		return
 	}
-	if foundItem == nil || foundItem.ID != id {
+	if foundItem == nil || foundItem.ID < 1 {
 		errCode = http.StatusNotFound
 		err = constants.Http404ErrorMessage(MODEL_NAME)
 		return
 	}
 
+	// Check school
+	newRequest := *request
+	if ctxData.Jwt.SchoolID > 0 {
+		newRequest.SchoolID = ctxData.Jwt.SchoolID
+	}
+
 	// Format request
 	item := &model.TeacherClassSubjectUnit{
-		TeacherID:      request.TeacherID,
-		YearID:         request.YearID,
-		ClassSubjectID: request.ClassSubjectID,
-		UnitID:         request.UnitID,
+		TeacherID:      newRequest.TeacherID,
+		YearID:         newRequest.YearID,
+		ClassSubjectID: newRequest.ClassSubjectID,
+		UnitID:         newRequest.UnitID,
 	}
 
 	// Check unique
@@ -378,9 +407,28 @@ func (service *Service) UpdateTeacherClassSubjectUnit(
 }
 
 func (service *Service) Delete(
-	inputJwtToken *types.JwtToken,
+	ctxData *types.ContextData,
 	id int64,
 ) (affectedRows int64, errCode int, err error) {
+	// Check if the item exists
+	var foundItem *model.Teacher
+	if ctxData.User.Feature != constants.FeatureAdmin {
+		foundItem, err = service.Repository.GetByIDSchoolID(id, ctxData.Jwt.SchoolID)
+	} else {
+		foundItem, err = service.Repository.GetByID(id)
+	}
+	if err != nil {
+		errCode = http.StatusInternalServerError
+		err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
+		return
+	}
+	if foundItem == nil || foundItem.ID < 1 {
+		errCode = http.StatusNotFound
+		err = constants.Http404ErrorMessage(MODEL_NAME)
+		return
+	}
+
+	// Delete
 	affectedRows, err = service.Repository.DeleteByID(id)
 	if err != nil {
 		errCode = http.StatusInternalServerError
@@ -396,7 +444,7 @@ func (service *Service) Delete(
 }
 
 func (service *Service) DeleteTeacherClassSubjectUnit(
-	inputJwtToken *types.JwtToken,
+	ctxData *types.ContextData,
 	id int64,
 ) (affectedRows int64, errCode int, err error) {
 	affectedRows, err = service.Repository.DeleteTeacherClassSubjectUnitByID(id)
@@ -413,8 +461,17 @@ func (service *Service) DeleteTeacherClassSubjectUnit(
 	return
 }
 
-func (service *Service) DeleteMultiple(inputJwtToken *types.JwtToken, list []int64) (affectedRows int64, errCode int, err error) {
-	affectedRows, err = service.Repository.DeleteMultipleByID(list)
+func (service *Service) DeleteMultiple(
+	ctxData *types.ContextData,
+	list []int64,
+) (affectedRows int64, errCode int, err error) {
+	// Check school
+	var foundSchool int64
+	if ctxData.Jwt.SchoolID > 0 {
+		foundSchool = ctxData.Jwt.SchoolID
+	}
+
+	affectedRows, err = service.Repository.DeleteMultipleByID(list, foundSchool)
 	if err != nil {
 		errCode = http.StatusInternalServerError
 		err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
@@ -428,8 +485,15 @@ func (service *Service) DeleteMultiple(inputJwtToken *types.JwtToken, list []int
 	return
 }
 
-func (service *Service) Get(inputJwtToken *types.JwtToken, id int64) (result *model.Teacher, errCode int, err error) {
-	result, err = service.Repository.GetByID(id)
+func (service *Service) Get(
+	ctxData *types.ContextData,
+	id int64,
+) (result *model.Teacher, errCode int, err error) {
+	if ctxData.User.Feature != constants.FeatureAdmin {
+		result, err = service.Repository.GetByIDSchoolID(id, ctxData.Jwt.SchoolID)
+	} else {
+		result, err = service.Repository.GetByID(id)
+	}
 	if err != nil {
 		errCode = http.StatusInternalServerError
 		err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
@@ -444,10 +508,14 @@ func (service *Service) Get(inputJwtToken *types.JwtToken, id int64) (result *mo
 }
 
 func (service *Service) GetTeacherClassSubjectUnit(
-	inputJwtToken *types.JwtToken,
+	ctxData *types.ContextData,
 	id int64,
 ) (result *model.TeacherClassSubjectUnit, errCode int, err error) {
-	result, err = service.Repository.GetTeacherClassSubjectUnitByID(id)
+	if ctxData.User.Feature != constants.FeatureAdmin {
+		result, err = service.Repository.GetTeacherClassSubjectUnitByIDShoolID(id, ctxData.Jwt.SchoolID)
+	} else {
+		result, err = service.Repository.GetTeacherClassSubjectUnitByID(id)
+	}
 	if err != nil {
 		errCode = http.StatusInternalServerError
 		err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
@@ -462,12 +530,19 @@ func (service *Service) GetTeacherClassSubjectUnit(
 }
 
 func (service *Service) GetAll(
-	inputJwtToken *types.JwtToken,
+	ctxData *types.ContextData,
 	filter *types.Filter,
 	pagination *types.Pagination,
-	schoolID int64,
+	request *data.GetAllRequest,
 ) (result []model.Teacher, errCode int, err error) {
-	result, err = service.Repository.GetAll(filter, pagination, &data.GetAllRequest{SchoolID: schoolID})
+	// Check school
+	newRequest := *request
+	if ctxData.Jwt.SchoolID > 0 {
+		newRequest.SchoolID = ctxData.Jwt.SchoolID
+	}
+
+	// Get
+	result, err = service.Repository.GetAll(filter, pagination, &newRequest)
 	if err != nil {
 		errCode = http.StatusInternalServerError
 		err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
@@ -476,12 +551,19 @@ func (service *Service) GetAll(
 }
 
 func (service *Service) GetAllTeacherClassSubjectUnit(
-	inputJwtToken *types.JwtToken,
+	ctxData *types.ContextData,
 	filter *types.Filter,
 	pagination *types.Pagination,
 	request *data.GetAllTeacherClassSubjectUnitRequest,
 ) (result []model.TeacherClassSubjectUnit, errCode int, err error) {
-	result, err = service.Repository.GetAllTeacherClassSubjectUnit(filter, pagination, request)
+	// Check school
+	newRequest := *request
+	if ctxData.Jwt.SchoolID > 0 {
+		newRequest.SchoolID = ctxData.Jwt.SchoolID
+	}
+
+	// Get
+	result, err = service.Repository.GetAllTeacherClassSubjectUnit(filter, pagination, &newRequest)
 	if err != nil {
 		errCode = http.StatusInternalServerError
 		err = constants.Http500ErrorMessage(DEFAULT_ERROR_MESSAGE)
